Move repl-session module to nrepl directory
