+++ src/debugger/decorations.ts
@@ -6,7 +6,7 @@ import { NReplSession } from '../nrepl';
 import * as util from '../utilities';
 import lsp from '../lsp';
 import { getStateValue } from '../../out/cljs-lib/cljs-lib';
-import * as replSession from '../repl-session';
+import * as replSession from '../nrepl/repl-session';
 
 let enabled = false;
