+++ src/namespace.ts
@@ -5,7 +5,7 @@ import { LispTokenCursor } from './cursor-doc/token-cursor';
 import { Token } from './cursor-doc/clojure-lexer';
 import * as outputWindow from './results-output/results-doc'
 import * as utilities from './utilities';
-import * as replSession from './repl-session';
+import * as replSession from './nrepl/repl-session';
 
 export function getNamespace(doc: vscode.TextDocument) {
     if (outputWindow.isResultsDoc(doc)) {
