+++ src/nrepl/repl-start.ts
@@ -7,7 +7,7 @@ import * as sequence from "./connectSequence";
 import * as jackIn from "./jack-in";
 import * as outputWindow from '../results-output/results-doc';
 import { getWorkspaceConfig } from '../config';
-import * as replSession from '../repl-session';
+import * as replSession from './repl-session';
 
 export const USER_TEMPLATE_FILE_NAMES = ['user.clj'];
 export const HELLO_TEMPLATE_FILE_NAMES = ['hello_repl.clj', 'hello_paredit.clj', 'hello_clojure.clj'];
